- start moving plots from ohw into plotfunctions
