+++ libraries/OHW.py
@@ -136,6 +136,7 @@ class OHW():
         self.absMotions = np.sqrt(self.unitMVs[:,0]*self.unitMVs[:,0] + self.unitMVs[:,1]*self.unitMVs[:,1])# get absolute motions per frame        
     
         self.get_mean_absMotion()
+        self.prepare_quiver_components()
         self.calc_TimeAveragedMotion()
     
     def get_mean_absMotion(self):
@@ -154,7 +155,23 @@ class OHW():
         self.analysis_meta["results_folder"].mkdir(parents = True, exist_ok = True) #create folder for results if it doesn't exist
         self.timeindex = (np.arange(self.mean_absMotions.shape[0]) / self.videometa["fps"]).round(2)
         np.save(str(self.analysis_meta["results_folder"] / 'beating_kinetics.npy'), np.array([self.timeindex,self.mean_absMotions]))
+    
+    def prepare_quiver_components(self):
+        '''
+            sets all 0-motions to nan such that they won't be plotted in quiver plot
+            creates grid of coordinates
+        '''
         
+        self.MV_zerofiltered = Filters.zeromotion_to_nan(self.unitMVs, copy=True)
+        self.MotionX = self.MV_zerofiltered[:,0,:,:]
+        self.MotionY = self.MV_zerofiltered[:,1,:,:]
+        
+        bw = self.analysis_meta["MV_parameters"]["blockwidth"]
+        
+        self.MotionCoordinatesX, self.MotionCoordinatesY = np.meshgrid(
+            np.arange(bw/2, self.analysisImageStack.shape[2]-bw/2, bw)+1, 
+            np.arange(bw/2, self.analysisImageStack.shape[1]-bw/2+1, bw))  #changed arange range, double check!.. inconsistency in x and y?
+            
     def save_MVs(self):
         """
             saves raw MVs as npy file
@@ -170,192 +187,20 @@ class OHW():
 
     #def plot_scalebar(self):
     # moved to module: helpfunctions.insert_scalebar(imageStack, videometa, analysis_meta)
-             
-    def save_heatmap(self, singleframe = False, keyword = None, subfolder = None, *args, **kwargs):
-        """
-            saves either the selected frame (singleframe = framenumber) or the whole heatmap video (=False)
-        """
-        
-        #prepare figure
-        savefig_heatmaps, saveax_heatmaps = plt.subplots(1,1)
-      #  w,h = helpfunctions.get_figure_size(self.scaledImageStack[0],initial_val=12)
-      #  savefig_heatmaps.set_size_inches(w,h)
-        savefig_heatmaps.set_size_inches(16,12) 
-        saveax_heatmaps.axis('off')   
-        
-        scale_max = self.get_scale_maxMotion()
-        imshow_heatmaps = saveax_heatmaps.imshow(self.absMotions[0], vmin = 0, vmax = scale_max, cmap = "jet", interpolation="bilinear")#  cmap="inferno"
-        
-        cbar_heatmaps = savefig_heatmaps.colorbar(imshow_heatmaps)
-        cbar_heatmaps.ax.tick_params(labelsize=20)
-        for l in cbar_heatmaps.ax.yaxis.get_ticklabels():
-            l.set_weight("bold")
-        saveax_heatmaps.set_title('Motion [µm/s]', fontsize = 16, fontweight = 'bold')
-        
-        if keyword == None:
-            path_heatmaps = self.analysis_meta["results_folder"]/ "heatmap_results"
-        elif keyword == 'batch':
-            path_heatmaps = subfolder / "heatmap_results"
-        path_heatmaps.mkdir(parents = True, exist_ok = True) #create folder for results
-        
-        if singleframe != False:
-        # save only specified frame
-            imshow_heatmaps.set_data(self.absMotions[singleframe])
-            
-            heatmap_filename = str(path_heatmaps / ('heatmap_frame' + str(singleframe) + '.png'))
-            savefig_heatmaps.savefig(heatmap_filename,bbox_inches ="tight",pad_inches =0)
-        
-        else:
-        # save video
-            def make_frame_mpl(t):
 
-                frame = int(round(t*self.videometa["fps"]))
-                imshow_heatmaps.set_data(self.absMotions[frame])
-
-                return mplfig_to_npimage(savefig_heatmaps) # RGB image of the figure
-            
-            heatmap_filename = str(path_heatmaps / 'heatmapvideo.mp4')
-            duration = 1/self.videometa["fps"] * self.absMotions.shape[0]
-            animation = mpy.VideoClip(make_frame_mpl, duration=duration)
-         #   animation.resize((1500,800))
-            animation.write_videofile(heatmap_filename, fps=self.videometa["fps"])    
-
+    def save_heatmap(self, singleframe):
+        savepath = self.analysis_meta["results_folder"]/'heatmap_results'
+        plotfunctions.save_heatmap(ohw_dataset = self, savepath = savepath, singleframe=False)
+    
     def save_heatmap_thread(self, singleframe):
-        self.thread_save_heatmap = helpfunctions.turn_function_into_thread(self.save_heatmap, singleframe=False)
+        savepath = self.analysis_meta["results_folder"]/'heatmap_results'
+        self.thread_save_heatmap = helpfunctions.turn_function_into_thread(
+            plotfunctions.save_heatmap, ohw_dataset = self, savepath = savepath, singleframe=False)
         return self.thread_save_heatmap
-    
-    def save_quivervid3(self, singleframe = False, skipquivers = 1, t_cut = 0, *args, **kwargs):
-        """
-            saves a video with the normal beating, beating + quivers and velocity trace
-            or a single frame with the same three views
-        """
-        
-        scale_max = self.get_scale_maxMotion2()   
-        self.MV_zerofiltered = Filters.zeromotion_to_nan(self.unitMVs, copy=True)
-        self.MV_cutoff = Filters.cutoffMVs(self.MV_zerofiltered, max_length = scale_max, copy=True)
-        
-        self.MotionX = self.MV_cutoff[:,0,:,:]
-        self.MotionY = self.MV_cutoff[:,1,:,:]
-
-        blockwidth = self.MV_parameters["blockwidth"]
-        self.MotionCoordinatesX, self.MotionCoordinatesY = np.meshgrid(np.arange(blockwidth/2, self.scaledImageStack.shape[2], blockwidth), np.arange(blockwidth/2, self.scaledImageStack.shape[1], blockwidth))        
-           
-        #prepare figure
-        outputfigure = plt.figure(figsize=(14,10), dpi = 150)#figsize=(6.5,12)
-
-        gs = gridspec.GridSpec(3,2, figure=outputfigure)
-       # add_gridspec(3, 2)
-        gs.tight_layout(outputfigure)
-        plt.tight_layout()
-        
-        saveax_video = outputfigure.add_subplot(gs[0:2, 0])
-        saveax_video.axis('off')        
-        
-        saveax_quivers = outputfigure.add_subplot(gs[0:2, 1])
-        saveax_quivers.axis('off')
-
-        saveax_trace = outputfigure.add_subplot(gs[2,:])
-        saveax_trace.plot(self.timeindex, self.mean_absMotions, '-', linewidth = 2)
-        
-        saveax_trace.set_xlim(left = 0, right = self.timeindex[-1])
-        saveax_trace.set_ylim(bottom = 0)
-        saveax_trace.set_xlabel('t [s]', fontsize = 22)
-        saveax_trace.set_ylabel(u'$\mathrm{\overline {v}}$ [\xb5m/s]', fontsize = 22)
-        saveax_trace.tick_params(labelsize = 20)
-
-        for side in ['top','right','bottom','left']:
-            saveax_trace.spines[side].set_linewidth(2) 
-        
-        self.marker, = saveax_trace.plot(self.timeindex[0],self.mean_absMotions[0],'ro')
-        
-        #savefig_quivers, saveax_quivers = plt.subplots(1,1)
-        #savefig_quivers.set_size_inches(8, 10)
-        #saveax_quivers.axis('off')   
-    
-        ###### prepare video axis
-        imshow_video = saveax_video.imshow(self.scaledImageStack[0], vmin = self.videometa["Blackval"], vmax = self.videometa["Whiteval"], cmap = "gray")
-        
-        qslice=(slice(None,None,skipquivers),slice(None,None,skipquivers))
-        distance_between_arrows = blockwidth * skipquivers
-        arrowscale = 1 / (distance_between_arrows / scale_max)
-               
-        imshow_quivers = saveax_quivers.imshow(self.scaledImageStack[0], vmin = self.videometa["Blackval"], vmax = self.videometa["Whiteval"], cmap = "gray")
-        # adjust desired quiver plotstyles here!
-        quiver_quivers = saveax_quivers.quiver(self.MotionCoordinatesX[qslice], self.MotionCoordinatesY[qslice], self.MotionX[0][qslice], self.MotionY[0][qslice], pivot='mid', color='r', units ="xy", scale_units = "xy", angles = "xy", scale = arrowscale,  width = 4, headwidth = 3, headlength = 5, headaxislength = 5, minshaft =1.5) #width = 4, headwidth = 2, headlength = 3
-        #saveax_quivers.set_title('Motion [µm/s]', fontsize = 16, fontweight = 'bold')
-
-        path_quivers = self.analysis_meta["results_folder"] / "quiver_results"
-        path_quivers.mkdir(parents = True, exist_ok = True) #create folder for results
-
-        # parameters for cropping white border in output video
-        sizex, sizey = outputfigure.get_size_inches()*outputfigure.dpi
-        bbox = outputfigure.get_tightbbox(outputfigure.canvas.get_renderer())
-        bbox_bounds_px = np.round(np.asarray(bbox.extents*outputfigure.dpi)).astype(int)
-
-        # to do: introduce min/max to be on the safe side!
-        # reverse for np indexing
-        bbox_bounds_px[3] = sizey - bbox_bounds_px[1]#y1
-        bbox_bounds_px[1] = sizey - bbox_bounds_px[3]#y0
-
-        bbox_bounds_px[2] = sizex - bbox_bounds_px[0]#x1
-        bbox_bounds_px[0] = sizex - bbox_bounds_px[2]#x0
-
-        # save only specified frame       
-        if not isinstance(singleframe, bool):
-            imshow_quivers.set_data(self.scaledImageStack[singleframe])
-            imshow_video.set_data(self.scaledImageStack[singleframe])
-            quiver_quivers.set_UVC(self.MotionX[singleframe][qslice], self.MotionY[singleframe][qslice])
-            
-            self.marker.remove()
-            self.marker, = saveax_trace.plot(self.timeindex[singleframe],self.mean_absMotions[singleframe],'ro')
-            self.marker.set_clip_on(False)
-                
-            outputfigure.savefig(str(path_quivers / ('quiver3_frame' + str(singleframe) + '.png')))
-                  
-        else:
-            # save video
-            def make_frame_mpl(t):
-                #calculate the current frame number:
-                frame = int(round(t*self.videometa["fps"]))
-                
- #               if len(self.timeindex) > frame:
- #                   print('Frame: {}, Timeindex: {}'.format(frame, self.timeindex[frame]))
-                imshow_quivers.set_data(self.scaledImageStack[frame])
-                imshow_video.set_data(self.scaledImageStack[frame])
-                
-                try:
-                    quiver_quivers.set_UVC(self.MotionX[frame][qslice], self.MotionY[frame][qslice])
-                    
-                    self.marker.remove()
-                    self.marker, = saveax_trace.plot(self.timeindex[frame],self.mean_absMotions[frame],'ro')
-                    self.marker.set_clip_on(False)
-                except Exception:
-                    pass
-                
-                return mplfig_to_npimage(outputfigure)[bbox_bounds_px[1]:bbox_bounds_px[3],bbox_bounds_px[0]:bbox_bounds_px[2]] # RGB image of the figure  #150:1450,100:1950
-                
- #               else:
- #                   return
-                # slicing here really hacky! find better solution!
-                # find equivalent to bbox_inches='tight' in savefig
-                # mplfig_to_npimage just uses barer canvas.tostring_rgb()
-                # -> check how bbox_inches works under the hood
-                # -> in print_figure:
-                # if bbox_inches:
-                # call adjust_bbox to save only the given area
-            
-            quivers_filename = str(path_quivers / 'quivervideo3.mp4')
-            duration = 1/self.videometa["fps"] * self.MotionX.shape[0]
-            animation = mpy.VideoClip(make_frame_mpl, duration=duration)
-            
-            #cut clip if desired by user
-            animation_to_save = self.cut_clip(clip_full=animation, t_cut=t_cut)
-           
-            animation_to_save.write_videofile(quivers_filename, fps=self.videometa["fps"])
-            
 
     def save_quivervid3_thread(self, skipquivers, t_cut):
-        self.thread_save_quivervid3 = helpfunctions.turn_function_into_thread(self.save_quivervid3, singleframe = False, skipquivers=skipquivers, t_cut=t_cut)
+        self.thread_save_quivervid3 = helpfunctions.turn_function_into_thread(
+            self.save_quivervid3, singleframe = False, skipquivers=skipquivers, t_cut=t_cut)
         return self.thread_save_quivervid3
     
     def save_quiver(self, singleframe = False, skipquivers = 1, t_cut = 0, keyword = None, subfolder = None, *args, **kwargs):
@@ -367,7 +212,7 @@ class OHW():
         """
         
         # prepare MVs... needs refactoring as it's done twice!
-        scale_max = self.get_scale_maxMotion2()   
+        scale_max = self.get_scale_maxMotion2(self.absMotions)   
         self.MV_zerofiltered = Filters.zeromotion_to_nan(self.unitMVs, copy=True)
         self.MV_cutoff = Filters.cutoffMVs(self.MV_zerofiltered, max_length = scale_max, copy=True)
         
@@ -456,32 +301,6 @@ class OHW():
         
         else:
             return clip_full
-    
-    def get_scale_maxMotion(self):
-        """
-            --> deprecated
-            returns maximum for scaling of heatmap + arrows in quiver
-            selects frame with max motion (= one pixel), picks 95th percentile in this frame as scale max
-            ... find better metric
-            #max_motion = self.mean_maxMotion
-            #scale_max = np.mean(self.OHW.absMotions)    #should be mean of 1d-array of max motions
-            #scale_max = np.mean(np.max(self.OHW.absMotions,axis=(1,2)))
-        """
-        
-        max_motion_framenr = np.argmax(self.mean_absMotions)
-        max_motion_frame = self.absMotions[max_motion_framenr]
-        scale_min, scale_maxMotion = np.percentile(max_motion_frame, (0.1, 95))
-        return scale_maxMotion
-    
-    def get_scale_maxMotion2(self):
-        """
-            returns maximum for scaling of heatmap + arrows in quiver
-            v2: takes all frames into account, select positions where motion > 0, max represents 80th percentile
-        """
-        
-        absMotions_flat = self.absMotions[self.absMotions>0].flatten()
-        scale_min, scale_maxMotion = np.percentile(absMotions_flat, (0, 80))   
-        return scale_maxMotion
     
     def detect_peaks(self, ratio, number_of_neighbours):
         """
